+++ packages/pluggable-widgets-tools/src/typings-generator/__tests__/outputs/metadata-attribute.ts
@@ -3,9 +3,9 @@ export const attributeMetaDataWebOutput = `/**
  * WARNING: All changes made to this file will be overwritten
  * @author Mendix Widgets Framework Team
  */
-import { CSSProperties } from "react";
 import { AttributeMetaData, ListValue } from "mendix";
 import { Big } from "big.js";
+import { CSSProperties } from "react";
 
 export interface MyWidgetContainerProps {
     name: string;